backend/compositeservices/negotiate/requirements.txt
flask
flask-cors
requests
python-dotenv
pika
